[FLINK-5727] [table] Unify some API of batch and stream TableEnvironment
